CurrentAnimatedSprite to own file
